refactored solid elastic storages
