+++ async_api/src/services/film.py
@@ -3,71 +3,12 @@ from uuid import UUID
 
 from db.elastic import get_elastic
 from db.redis import get_redis
-from elasticsearch import AsyncElasticsearch, NotFoundError
+
 from redis.asyncio import Redis
 from services.cache import RedisCache
 
 from fastapi import Depends
 
-
-class FilmStorage:
-    def __init__(self, elastic: AsyncElasticsearch):
-        self.elastic = elastic
-
-    async def search_data(self, query, page_number, page_size):
-        search_query = {"query_string": {"default_field": "title", "query": query}}
-        docs = await self.elastic.search(
-            index="movies",
-            body={
-                "_source": ["id", "title", "imdb_rating"],
-                "from": (page_number - 1) * page_size,
-                "size": page_size,
-                "query": search_query,
-            },
-            params={"filter_path": "hits.hits._source"},
-        )
-        if not docs:
-            return None
-        return [film["_source"] for film in docs["hits"]["hits"]]
-
-    async def get_data_by_id(self, id: UUID) -> dict:
-        try:
-            doc = await self.elastic.get(index="movies", id=id)
-        except NotFoundError:
-            return None
-        return doc["_source"]
-
-    async def get_data_list(
-        self, sort: str, genre: UUID, page_number: int, page_size: int
-    ) -> list[dict] | None:
-        if sort[0] == "-":
-            sort = {sort[1:]: "desc"}
-        else:
-            sort = {sort: "asc"}
-        filter_query = (
-            {"match_all": {}}
-            if genre is None
-            else {
-                "nested": {
-                    "path": "genre",
-                    "query": {"bool": {"must": {"match": {"genre.id": genre}}}},
-                }
-            }
-        )
-        docs = await self.elastic.search(
-            index="movies",
-            body={
-                "_source": ["id", "title", "imdb_rating"],
-                "sort": sort,
-                "from": (page_number - 1) * page_size,
-                "size": page_size,
-                "query": filter_query,
-            },
-            params={"filter_path": "hits.hits._source"},
-        )
-        if not docs:
-            return None
-        return [film["_source"] for film in docs["hits"]["hits"]]
 
 
 class FilmService:
